fix: adapt trajectory save to custom meshio h5 location writing

+++ graphphysics/training/lightning_module.py
@@ -205,13 +205,16 @@ class LightningModule(L.LightningModule):
         logger.info(
             f"Validation Trajectory {archive_filename.split('_')[-1]} saved at {save_dir}."
         )
-        # The H5 archive is systematically created in cwd, we just need to move it
-        shutil.move(
-            src=os.path.join(
-                os.getcwd(), os.path.split(f"{xdmf_filename.replace('xdmf', 'h5')}")[1]
-            ),
-            dst=f"{xdmf_filename.replace('xdmf', 'h5')}",
-        )
+        # The H5 archive is systematically created in cwd with original meshio library, we just need to move it
+        h5_filename = os.path.split(f"{xdmf_filename.replace('xdmf', 'h5')}")
+        if os.path.exists(os.path.join(os.getcwd(), h5_filename[1])):
+            shutil.move(
+                src=os.path.join(
+                    os.getcwd(),
+                    h5_filename[1],
+                ),
+                dst=h5_filename,
+            )
 
     def _reset_validation_trajectory(self):
         self.current_val_trajectory += 1
